+++ apps/modgpt/src/dev-utils/configs.ts
@@ -25,4 +25,4 @@ export const areClerkKeysSet = (
 };
 
 export const clerkApplied = areClerkKeysSet(environment);
-export const isDevelopment = NODE_ENV === "development";
+export const isDevelopment = false; //NODE_ENV === "development";
